fix: address PR review findings on the warm-up publication barrier
Copilot and claude[bot] reviewed #1520 (#1466's fix) and found two things.

The javadoc on Catalog#markUnpublishable linked
markUnpublishableDueToInvalidSchema(SchemaAlteringException), but the method
takes EvitaInvalidUsageException - a stale cross-reference from before that
parameter was widened, contradicting the paragraph right below it that
explains why the wider type is caught.

AttributeFilterAcceleratorRefusalTest#shouldRefuseAnAcceleratorDeclaredOnAnAttributeWithNoFilterIndex
reopened storage and asserted the persisted schema without asserting that the
refusal raised the barrier itself - exactly the gap the ADR's own Verification
section already named as "one edit away from mattering": if
SetAttributeSchemaAcceleratedMutation ever gained a pre-flight check (its
sibling CreateAttributeSchemaMutation already has one), the refusal would
leave the catalog untouched and the disk assertion would pass trivially.

The barrier-assertion machinery - assertRefusalRaisesTheBarrier,
subscribeToSettledState, SettledStateLatch - moves out of
WarmUpRefusedSchemaPersistenceTest into its own
CatalogUnpublishableBarrierAssertions, parameterized by catalog name instead
of a hardcoded constant, so both test classes share it. The flagged test now
asserts the barrier before checking disk state, and its reopen no longer
races the deactivation the way the general-purpose helper it calls
documents. In passing, replaced this test's other stale "last session that
closed successfully" phrasing with the same correction already made
elsewhere in the PR.

Ref: #1466
